+++ packages/twenty-server/src/engine/core-modules/file/controllers/file.controller.ts
@@ -22,7 +22,7 @@ import {
 import { FileApiExceptionFilter } from 'src/engine/core-modules/file/filters/file-api-exception.filter';
 import { FilePathGuard } from 'src/engine/core-modules/file/guards/file-path-guard';
 import { FileService } from 'src/engine/core-modules/file/services/file.service';
-import { extractFileInfoFromRequest } from 'src/engine/core-modules/file/file.utils';
+import { extractFileInfoFromRequest } from 'src/engine/core-modules/file/utils/extract-file-info-from-request.utils';
 
 @Controller('files')
 @UseFilters(FileApiExceptionFilter)
