+++ packages/twenty-server/src/engine/core-modules/file/utils/build-file-info.utils.ts
@@ -1,0 +1,13 @@
+import { v4 as uuidV4 } from 'uuid';
+
+export const buildFileInfo = (filename: string) => {
+  const parts = filename.split('.');
+
+  const ext = parts.length > 1 ? parts.pop() || '' : '';
+
+  const id = uuidV4();
+
+  const name = `${id}${ext ? `.${ext}` : ''}`;
+
+  return { ext, name, id };
+};
