+++ packages/twenty-server/src/engine/core-modules/file/utils/extract-file-id-from-path.utils.ts
@@ -1,3 +1,15 @@
+import { isNonEmptyString } from '@sniptt/guards';
+
 export const extractFileIdFromPath = (path: string) => {
-  return path.split('/').reverse()[0];
+  if (path.includes('/') && path.slice(-1) === '/') {
+    throw new Error(`Cannot extract id from folder path '${path}'`);
+  }
+
+  const fileId = path.split('/').reverse()[0];
+
+  if (!isNonEmptyString(fileId)) {
+    throw new Error(`Cannot extract id from empty path`);
+  }
+
+  return fileId;
 };
